added AR page into product detail page
