--- demo/lib/pages/ar_view_page.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ARViewPage extends StatefulWidget {
-  const ARViewPage({super.key});
-
-  @override
-  State<ARViewPage> createState() => _ARViewPageState();
-}
-
-class _ARViewPageState extends State<ARViewPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.black,
-        elevation: 0,
-        title: const Text("AR View", style: TextStyle(fontSize: 20)),
-      ),
-      body: const Center(
-        child: Text(
-          "AR View Page",
-          style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
-        ),
-      ),
-    );
-  }
-}
