+++ demo/lib/pages/home_page.dart
@@ -4,7 +4,7 @@ import '../widgets/bottom_navagiationbar.dart';
 import 'home_content.dart';
 
 import 'wishlist_page.dart';
-import 'ar_view_page.dart';
+import 'ar_screen.dart';
 import 'cart_page.dart';
 import 'custom_page.dart';
 
@@ -21,7 +21,7 @@ class _HomePageState extends State<HomePage> {
   final List<Widget> pages = const [
     HomeContent(),
     WishlistPage(),
-    ARViewPage(),
+    ARScreen(),
     CartPage(),
     CustomPage(),
   ];
